Split schemas into modules
